Moved user_settings to a separate module.
