+++ lib/modules/user_settings/user_settings.dart
@@ -1,0 +1,4 @@
+library user_settings;
+
+export 'src/services/i_settings_database.dart';
+export 'src/services/settings_database.dart';
